Moved some test specific classes to a new package.
Instead of having them in the test root, they are now instead in a new package "test".
