+++ MovieFinder/src/test/java/edu/chalmers/dat076/moviefinder/test/AbstractIntegrationTest.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package edu.chalmers.dat076.moviefinder;
+package edu.chalmers.dat076.moviefinder.test;
 
 import java.sql.Connection;
 import java.sql.SQLException;
